+++ execution_engine/src/core/tracking_copy/tests.rs
@@ -7,8 +7,8 @@ use casper_types::{
     account::{Account, AccountHash, AssociatedKeys, Weight, ACCOUNT_HASH_LENGTH},
     contracts::NamedKeys,
     gens::*,
-    AccessRights, CLValue, Contract, EntryPoints, HashAddr, Key, KeyTag, ProtocolVersion, URef,
-    U256, U512,
+    AccessRights, CLValue, Contract, EntryPoints, HashAddr, Key, KeyTag, ProtocolVersion,
+    StoredValue, URef, U256, U512,
 };
 
 use super::{
@@ -22,7 +22,6 @@ use crate::{
     },
     shared::{
         newtypes::{Blake2bHash, CorrelationId},
-        stored_value::{gens::stored_value_arb, StoredValue},
         transform::Transform,
     },
     storage::{