+++ node/src/components/consensus/protocols/highway/config.rs
@@ -16,14 +16,14 @@ pub struct Config {
     pub unit_hashes_folder: PathBuf,
     /// The duration for which incoming vertices with missing dependencies are kept in a queue.
     pub pending_vertex_timeout: TimeDiff,
-    /// If the current era's protocol state has not progressed for this long, request the latest
-    /// state from peers.
-    pub standstill_timeout: TimeDiff,
-    /// If another `standstill_timeout` passes assume we failed to join the network and restart.
-    #[serde(default = "default_shutdown_on_standstill")]
-    pub shutdown_on_standstill: bool,
+    /// If the initial era's protocol state has not progressed for this long, restart.
+    pub standstill_timeout: Option<TimeDiff>,
+    /// Request the latest protocol state from a random peer periodically, with this interval.
+    pub request_state_interval: Option<TimeDiff>,
     /// Log inactive or faulty validators periodically, with this interval.
-    pub log_participation_interval: TimeDiff,
+    pub log_participation_interval: Option<TimeDiff>,
+    /// Log synchronizer state periodically, with this interval.
+    pub log_synchronizer_interval: Option<TimeDiff>,
     /// Log the size of every incoming and outgoing serialized unit.
     pub log_unit_sizes: bool,
     /// The maximum number of blocks by which execution is allowed to lag behind finalization.
@@ -34,18 +34,15 @@ pub struct Config {
     pub round_success_meter: RSMConfig,
 }
 
-fn default_shutdown_on_standstill() -> bool {
-    false
-}
-
 impl Default for Config {
     fn default() -> Self {
         Config {
             unit_hashes_folder: Default::default(),
             pending_vertex_timeout: "10sec".parse().unwrap(),
-            standstill_timeout: "1min".parse().unwrap(),
-            shutdown_on_standstill: false,
-            log_participation_interval: "10sec".parse().unwrap(),
+            standstill_timeout: None,
+            request_state_interval: Some("10sec".parse().unwrap()),
+            log_participation_interval: Some("10sec".parse().unwrap()),
+            log_synchronizer_interval: Some("5sec".parse().unwrap()),
             log_unit_sizes: false,
             max_execution_delay: 3,
             max_requests_for_vertex: 5,